### 2021-12-01 #3971 @whedon
[PRE REVIEW]: helayo: Reconstructing Sanskrit texts from manuscript witnesses
**Submitting author:** @chchch (<a href="http://orcid.org/0000-0001-8247-9374">Charles Li</a>)
**Repository:** <a href="https://github.com/chchch/sanskrit-alignment/" target ="_blank">https://github.com/chchch/sanskrit-alignment/</a>
**Version:** v0.5 (helayo) v0.8 (matrix editor)
