### 2021-12-20 #3971 @chchch
Hello,

Starting from the bottom of the list, I suggested JacksonMaxfield or chrisbrickhouse, but I'm new to this so I'm not sure what the relevant criteria are. I know this submission might seem a bit out of the ordinary — Sanskrit philology in Haskell and Javascript!
