### 2021-12-21 #4022 @whedon
[REVIEW]: helayo: Reconstructing Sanskrit texts from manuscript witnesses
**Submitting author:** @chchch (<a href="http://orcid.org/0000-0001-8247-9374">Charles Li</a>)
**Repository:** https://github.com/chchch/sanskrit-alignment/
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.55
