### 2021-12-23 chchch/sanskrit-alignment#1 @kinow
Add a CONTRIBUTING.md file?
Hi @chchch

I'm reviewing the paper for sanskrit-alignment (helayo, and matrix-editor). Very interesting software!

See the [review issue](https://github.com/openjournals/joss-reviews/issues/4022) for what's being reviewed. This issue is about this item of the checklist:

>Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I think it's clear where users can report issues and seek support (over here, in the issues/pull requests sections I think).

However, there are no clear guidelines for third parties wishing to contribute to the software. Would you consider adding one? Regardless of the review requirements, I really think it could help other users wishing to contribute to your software in deciding whether to contribute or not.

- [Wrangling Web Contributions: How to Build a CONTRIBUTING.md](https://mozillascience.github.io/working-open-workshop/contributing/)
- [Adding a CONTRIBUTING file](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors#adding-a-contributing-file)

Cheers
-Bruno
