### 2021-12-23 chchch/sanskrit-alignment#2 @kinow
JavaScript errors when importing a tree
Hi @chchch

I used the web interface, matrix-editor, with the example XML [c3.2.1-0.xml](https://github.com/chchch/sanskrit-alignment/blob/ab1a9a5b492f1738f36a375150b4da3e014accd4/example/xml/c3.2.1-0.xml).

Then I loaded the tree XML [alternative-rooted-tree.xml](https://github.com/chchch/sanskrit-alignment/blob/ab1a9a5b492f1738f36a375150b4da3e014accd4/example/trees/alternative-rooted-tree.xml).

I simply picked the first file in the listing of XMLs, and ditto for the tree file. Without worrying about correctness. I expected the software would either validate the inputs and raise an error, or show me something.

This is about this item in the [review](https://github.com/openjournals/joss-reviews/issues/4022) checklist for the JOSS paper: "Functionality: Have the functional claims of the software been confirmed?".

Instead of seeing the tree as in the paper, all that I can see is the table, and some errors in the browser console.

![image](https://user-images.githubusercontent.com/304786/147204267-f136d6eb-aa44-4d2c-925e-976d475087c1.png)

-Bruno
