### 2021-12-23 chchch/sanskrit-alignment#3 @kinow
Add unit tests for JavaScript
I had some JS errors in #2, and while looking at this item in the [review checklist](https://github.com/openjournals/joss-reviews/issues/4022), “Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?”, I confirmed the Haskell part is being tested in GitHub actions, but the JavaScript appears to be lacking tests.

It also doesn't have package management, instead all dependencies in the `lib` folder. You could add a `package.json`, maybe use `webpack`, and then move the dependencies to `package.json`, and run mocha/chai/jest/etc for tests. Some functions like the one that merges objects can be replaced by calls to libraries like lodash, underscore, or ramda. But the tests are the most important, I think.

Cheers
-Bruno
