### 2021-12-23 #4022 @kinow
Hi @chchch

I've added a few issues in your sanskrit-alignment and linked to this issue. We must keep the discussion over there to avoid lengthy discussions on this issue :+1:

However, there is one review issue that isn't related to the code there, so I thought we could discuss it here.

>State of the field: Do the authors describe how this software compares to other commonly-used packages?

In the paper draft I believe you mention that while the techniques used in helayo is similar to those used in the bioinformatics field, but very different to modern bioinformatics techniques that are too specific to be used in natural language processing.

But reading the paper I had a few questions, like “could I use it for Chinese and Japanese manuscripts?”, I believe they suffer similar issue to what you described in the motivation section of the paper. Some papers include a section about future work, improvements, etc, not sure if necessary here (cc @galessiorob ), but looks like it could be useful for developers/linguists/etc interested in your work.

Thanks and great job on helayo and matrix-editor! :tada:
-Bruno
