### 2021-12-23 chchch/sanskrit-alignment#2 @chchch
Thanks, I just updated the example files, although I'm not sure why it would cause that error. From your screenshot, it looks like you're using an old version of the matrix editor somehow — are you using the version hosted on GitHub at https://chchch.github.io/sanskrit-alignment/matrix-editor/ ?
