+++ CONTRIBUTING.md
@@ -1,0 +1,17 @@
+# Contributing to the project
+
+Comments, suggestions, bug reports, and pull requests are welcome!
+
+## Reporting bugs
+
+The easiest way to report a bug is by creating a new issue in the GitHub repository. Check the existing list of issues to see if your bug has been reported; if not, go ahead and make a new one.
+
+When explaining your problem, providing the following details will help the maintainers in assessing the situation:
+
+* The exact steps which reproduce the problem
+* The behaviour observed after following the steps
+* The operating system and, if applicable, the web browser used
+
+## Pull requests
+
+The code in this project uses an indent of 4 spaces, across all programming languages. 
