### 2021-12-23 chchch/sanskrit-alignment#2 @chchch
Ah, my fault, I think you used the link from the README, which is outdated.
