### 2021-12-23 chchch/sanskrit-alignment#2 @kinow
> Ah, my fault, I think you used the link from the README, which is outdated.

Ah! I see. I have the matrix-editor locally too with `python -m http.server 8000` from its directory, but I didn't test trees with it.

Looks like it's working now! And since the link has been fixed, and the GH actions will test and then GH pages will be updated, this issue won't happen again for other users. Let's close it. Thanks @chchch !

![image](https://user-images.githubusercontent.com/304786/147295019-01c4c78e-0111-4037-b235-dfae5b86ebd3.png)
