### 2021-12-23 chchch/sanskrit-alignment#1 @kinow
@chchch I just rebased my local clone, and noticed you added a contributing file. I confirmed the file looks OK, and GH UI is already telling users about it :tada:

![image](https://user-images.githubusercontent.com/304786/147295199-829313de-9b5a-4609-ab95-efd00df6bf6f.png)

Feel free to close this issue if you'd like (not sure if you were planning on writing more in that file.)

Thanks
-Bruno
