### 2021-12-24 5db20d7
added details about licensing
