### 2021-12-24 chchch/sanskrit-alignment#1 @chchch
I thought the my contributing file was a little thin, and I was trying to think of what else I could add. I'll close this issue for now, but let me know if you have any suggestions!
