+++ paper.md
@@ -40,4 +40,10 @@ The `matrix editor` is written in Javascript and can be used either online or of
 
 ![The matrix editor.](paper-fig1.png)
 
+# Discussion
+
+Previous work on textual phylogeny has focused on texts in Western languages (e.g., [@Barbrook:1998]), and, consequently, the tools developed from those projects embody assumptions about how Western languages are spoken, written, and transmitted (e.g., [@collatex]), in the same way that tools produced for bioinformatics embody assumptions about biological sequences. While any set of sequences ― composed of text in various languages, of DNA, of proteins ― can be analyzed using the same fundamental algorithms, the results are meaningless unless we use domain-specific knowledge to refine those algorithms and to interpret those results. 
+
+Although `helayo` has been designed specifically to align Sanskrit texts, each step of the process — normalization, tokenization, and sequence alignment — is architecturally distinct, and can be modified to work with other languages. Previously, textual scholars created alignments (or "collations") by hand, comparing different manuscripts to a reference text and using their own judgement to note down the most important differences. The process was slow, but, more importantly, it was not reproducible. By using formal algorithms to align texts, and, furthermore, by refining those algorithms to work with texts from a specific language and tradition, we can begin to create reproducible, testable models of how textual transmission actually works.
+
 # References
