+++ paper.bib
@@ -51,6 +51,15 @@ pages = {406-316},
 year= 1971
 }
 
+@article {Barbrook:1998,
+author = {{Barbrook}, Adrian C., {Howe}, Christopher J., {Blake}, Norman, {Robinson}, Peter}
+title = "The phylogeny of The Canterbury Tales",
+journal = "Nature",
+number = {394},
+pages = {839},
+year = 1998
+}
+
 @misc {align-affine,
 author = {{Li}, Charles},
 title = {align-affine: Sequence alignment with an affine gap penalty model},
@@ -58,4 +67,13 @@ year = 2021,
 publisher = {GitHub},
 journal = {GitHub repository},
 url = {https://github.com/chchch/align-affine}
+}
+
+@misc {collatex,
+author = {{Dekker}, Ronald Haentjens, {Middell}, Gregor},
+title = {CollateX — Software for Collating Textual Sources},
+year = 2019,
+publisher = {GitHub},
+journal = {GitHub repository},
+url = {https://github.com/interedition/collatex}
 }
