+++ paper.bib
@@ -52,7 +52,7 @@ year= 1971
 }
 
 @article {Barbrook:1998,
-author = {{Barbrook}, Adrian C., {Howe}, Christopher J., {Blake}, Norman, {Robinson}, Peter}
+author = {{Barbrook}, Adrian C., {Howe}, Christopher J., {Blake}, Norman, {Robinson}, Peter},
 title = "The phylogeny of The Canterbury Tales",
 journal = "Nature",
 number = {394},
