+++ paper.bib
@@ -52,7 +52,7 @@ year= 1971
 }
 
 @article {Barbrook:1998,
-author = {{Barbrook}, Adrian C., {Howe}, Christopher J., {Blake}, Norman, {Robinson}, Peter},
+author = {{Barbrook}, Adrian C. and {Howe}, Christopher J. and {Blake}, Norman and {Robinson}, Peter},
 title = "The phylogeny of The Canterbury Tales",
 journal = "Nature",
 number = {394},
@@ -70,7 +70,7 @@ url = {https://github.com/chchch/align-affine}
 }
 
 @misc {collatex,
-author = {{Dekker}, Ronald Haentjens, {Middell}, Gregor},
+author = {{Dekker}, Ronald Haentjens and {Middell}, Gregor},
 title = {CollateX — Software for Collating Textual Sources},
 year = 2019,
 publisher = {GitHub},
