+++ paper.bib
@@ -53,7 +53,7 @@ year= 1971
 
 @article {Barbrook:1998,
 author = {{Barbrook}, Adrian C. and {Howe}, Christopher J. and {Blake}, Norman and {Robinson}, Peter},
-title = "The phylogeny of The Canterbury Tales",
+title = "The phylogeny of The {Canterbury Tales}",
 journal = "Nature",
 number = {394},
 pages = {839},
