### 2021-12-26 #4022 @chchch
Really sorry about the issue spam — I finally learned to use the preview service to get it right first!

I've added a discussion section at the end, to talk about working with different languages, and what it means to create alignments of different kinds of sequences. I tried to keep it focused on the software, but I had to bring in a bit of theory — I'm still not totally sure if I have the tone right for a JOSS paper, though. Let me know if it makes sense and if it answers your questions!
