### 2021-12-27 9d9e6eb
WIP adding package.json and adding a sample test
