+++ matrix-editor/lib/sanscript.js
@@ -7,10 +7,8 @@
  * Released under the MIT and GPL Licenses.
  */
 
-(function(Sanscript) {
-    'use strict';
-
-    Sanscript.defaults = {
+export class Sanscript {
+    defaults = {
         skip_sgml: false,
         syncope: false
     };
@@ -29,401 +27,401 @@
      * (Since "abugida" is not a well-known term, Sanscript uses "Brahmic"
      * and "roman" for clarity.)
      */
-    var schemes = Sanscript.schemes = {
-            balinese: {
-                vowels: 'ᬅ ᬆ ᬇ ᬈ ᬉ ᬊ ᬋ ᬌ ᬍ ᬎ  ᬏ ᬐ  ᬑ ᬒ'.split(' '),
+    schemes = {
+        balinese: {
+            vowels: 'ᬅ ᬆ ᬇ ᬈ ᬉ ᬊ ᬋ ᬌ ᬍ ᬎ  ᬏ ᬐ  ᬑ ᬒ'.split(' '),
 
 
-                vowel_marks: 'ᬵ ᬶ ᬷ ᬸ ᬹ ᬺ ᬻ ᬼ ᬽ ᭂ ᬾ ᬿ ᭃ ᭀ ᭁ'.split(' '),
+            vowel_marks: 'ᬵ ᬶ ᬷ ᬸ ᬹ ᬺ ᬻ ᬼ ᬽ ᭂ ᬾ ᬿ ᭃ ᭀ ᭁ'.split(' '),
 
-                other_marks: ['ᬂ','ᬄ','ᬁ',''],
-                virama: ['᭄'],
+            other_marks: ['ᬂ','ᬄ','ᬁ',''],
+            virama: ['᭄'],
 
-                consonants: 'ᬓ ᬔ ᬕ ᬖ ᬗ ᬘ ᬙ ᬚ ᬛ ᬜ ᬝ ᬞ ᬟ ᬠ ᬡ ᬢ ᬣ ᬤ ᬥ ᬦ ᬧ ᬨ ᬩ ᬪ ᬫ ᬬ ᬭ ᬮ ᬯ ᬰ ᬱ ᬲ ᬳ'.split(' '),
+            consonants: 'ᬓ ᬔ ᬕ ᬖ ᬗ ᬘ ᬙ ᬚ ᬛ ᬜ ᬝ ᬞ ᬟ ᬠ ᬡ ᬢ ᬣ ᬤ ᬥ ᬦ ᬧ ᬨ ᬩ ᬪ ᬫ ᬬ ᬭ ᬮ ᬯ ᬰ ᬱ ᬲ ᬳ'.split(' '),
 
-                symbols: '᭐ ᭑ ᭒ ᭓ ᭔ ᭕ ᭖ ᭗ ᭘ ᭙   ᭞ ᭟'.split(' '),
+            symbols: '᭐ ᭑ ᭒ ᭓ ᭔ ᭕ ᭖ ᭗ ᭘ ᭙   ᭞ ᭟'.split(' '),
 
 
-                zwj: ['\u200D'],
+            zwj: ['\u200D'],
 
-                skip: [''],
+            skip: [''],
 
-            },
-            /* Bengali
-         * -------
-         * 'va' and 'ba' are both rendered as ব.
-         */
-            bengali: {
-                vowels: 'অ আ ই ঈ উ ঊ ঋ ৠ ঌ ৡ  এ ঐ  ও ঔ'.split(' '),
-                vowel_marks: 'া ি ী ু ূ ৃ ৄ ৢ ৣ  ে ৈ  ো ৌ'.split(' '),
-                other_marks: 'ং ঃ ঁ'.split(' '),
-                virama: ['্'],
[…]
