+++ matrix-editor/lib/script.js
@@ -1,19 +1,24 @@
+import {Sanscript} from "./sanscript.js";
+import {Normalizer} from "./normalize.js";
+import * as xsltsheet from "./xsltsheet.js";
+import * as xsltsheetTree from "./xsltsheet-tree.js";
+
 const comboView = (function() {
 
     'use strict';
-    
-    const Sanscript = window.Sanscript;
+
+    const sanscript = new Sanscript();
     const Smits = window.Smits;
     const FileSaver = window.FileSaver;
     const CSV = window.CSV;
-    const lgXSLT = window.lgXSLT;
-    const csvXSLT = window.csvXSLT;
-    const prettyXSLT = window.prettyXSLT;
-    const lemmaXSLT = window.lemmaXSLT;
-    const treeXSLT = window.treeXSLT;
-    const matrixXSLT = window.matrixXSLT;
-    const Normalizer = window.Normalizer;
-    
+    const lgXSLT = xsltsheetTree.lgXSLT;
+    const csvXSLT = xsltsheet.csvXSLT;
+    const prettyXSLT = xsltsheet.prettyXSLT;
+    const lemmaXSLT = xsltsheet.lemmaXSLT;
+    const treeXSLT = xsltsheetTree.treeXSLT;
+    const matrixXSLT = xsltsheet.matrixXSLT;
+    const normalizer = new Normalizer(sanscript);
+
     var _filename;
     var _xml;
     const _teins = 'http://www.tei-c.org/ns/1.0';
@@ -107,10 +112,10 @@ const comboView = (function() {
 
         smush: function(text,placeholder) {
             return text.toLowerCase()
-        
+
             // remove space between a word that ends in a consonant and a word that begins with a vowel
                 .replace(/([ḍdrmvynhs]) ([aāiīuūṛeoéó])/g, '$1$2'+placeholder)
-        
+
             // remove space between a word that ends in a consonant and a word that begins with a consonant
                 .replace(/([kgcjṭḍtdpb]) h/g, '$1\u200C'+placeholder+'h')
                 .replace(/([kgcjñḍtdnpbmrlyvśṣsṙ]) ([kgcjṭḍtdnpbmyrlvśṣshḻ])/g, '$1'+placeholder+'$2')
@@ -149,12 +154,12 @@ const comboView = (function() {
 
             return in_deva.replace(/¯/g, 'ꣻ');
         },
-    
+
         malayalam: function(text,p) {
 
             const placeholder = p || '';
             const options = {skip_sgml: true};
-	
+
             const chillu = {
                 'ക':'ൿ',
                 'ത':'ൽ',
@@ -173,21 +178,21 @@ const comboView = (function() {
                 .replace(/ḿ/g,'ṃ')
                 .replace(/í/g,'i') // no pṛṣṭhamātras
                 .replace(/ú/g,'u')
-                .replace(/é/g,'e'); 
+                .replace(/é/g,'e');
 
             const in_mlym = Sanscript.t(smushed,'iast','malayalam',options);
-	
+
             // use dot reph
             return in_mlym.replace(/(^|[^്])ര്(?=\S)/g,'$1ൎ')
-        
-            // use chillu final consonants	
-                .replace(/([കതനമര])്(?![^\s\u200C,—’―])/g, 
+
+            // use chillu final consonants
+                .replace(/([കതനമര])്(?![^\s\u200C,—’―])/g,
                     function(match,p1) {
                         return chillu[p1];
                     }
                 );
         },
-    
+
         grantha: function(text,p) {
 
             const placeholder = p || '';
@@ -201,8 +206,8 @@ const comboView = (function() {
                 ['ശ','\ue1d2'],
             ]);
 
[…]
