### 2021-12-27 #4022 @kinow
@galessiorob I think I'm done with the review for now. The paper is short but very well written (great job @chchch :clap: ).

The code repository is also well organized, and it's easy to locate the parts of the software mentioned in the paper. There are no performance claims in the paper, so I left the box unchecked.

I reviewed all the other items. The Automated Tests part, I think it would be an OK as per the review criteria for Tests: https://joss.readthedocs.io/en/latest/review_criteria.html#tests

The `helayo` software has tests, that are run with GH Actions. However, the other part of the system, `matrix-editor`, doesn't have tests. I created an issue for that [here](https://github.com/chchch/sanskrit-alignment/issues/3) and an example [PR](https://github.com/chchch/sanskrit-alignment/pull/5) showing one way to add tests for the JavaScript code.

There are a few issues with the dependencies used in `matrix-editor` too (added a comment to the PR above about it).

It's possible to use the `matrix-editor` as-is, following the paper instructions. So I per the text in the review criteria, I think my review is:

>OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)

If that's good enough, I can check the checkbox and finalize my review. What do you think?

Thank you!
-Bruno
