+++ matrix-editor/lib/sanscript.js
@@ -1,812 +1,0 @@
-/**
- * Sanscript
- *
- * Sanscript is a Sanskrit transliteration library. Currently, it supports
- * other Indian languages only incidentally.
- *
- * Released under the MIT and GPL Licenses.
- */
-
-(function(Sanscript) {
-    'use strict';
-
-    Sanscript.defaults = {
-        skip_sgml: false,
-        syncope: false
-    };
-
-    /* Schemes
-     * =======
-     * Schemes are of two kinds: "Brahmic" and "roman." "Brahmic" schemes
-     * describe abugida scripts found in India. "Roman" schemes describe
-     * manufactured alphabets that are meant to describe or encode Brahmi
-     * scripts. Abugidas and alphabets are processed by separate algorithms
-     * because of the unique difficulties involved with each.
-     *
-     * Brahmic consonants are stated without a virama. Roman consonants are
-     * stated without the vowel 'a'.
-     *
-     * (Since "abugida" is not a well-known term, Sanscript uses "Brahmic"
-     * and "roman" for clarity.)
-     */
-    var schemes = Sanscript.schemes = {
-            balinese: {
-                vowels: 'ᬅ ᬆ ᬇ ᬈ ᬉ ᬊ ᬋ ᬌ ᬍ ᬎ  ᬏ ᬐ  ᬑ ᬒ'.split(' '),
-
-
-                vowel_marks: 'ᬵ ᬶ ᬷ ᬸ ᬹ ᬺ ᬻ ᬼ ᬽ ᭂ ᬾ ᬿ ᭃ ᭀ ᭁ'.split(' '),
-
-                other_marks: ['ᬂ','ᬄ','ᬁ',''],
-                virama: ['᭄'],
-
-                consonants: 'ᬓ ᬔ ᬕ ᬖ ᬗ ᬘ ᬙ ᬚ ᬛ ᬜ ᬝ ᬞ ᬟ ᬠ ᬡ ᬢ ᬣ ᬤ ᬥ ᬦ ᬧ ᬨ ᬩ ᬪ ᬫ ᬬ ᬭ ᬮ ᬯ ᬰ ᬱ ᬲ ᬳ'.split(' '),
-
-                symbols: '᭐ ᭑ ᭒ ᭓ ᭔ ᭕ ᭖ ᭗ ᭘ ᭙   ᭞ ᭟'.split(' '),
-
-
-                zwj: ['\u200D'],
-
-                skip: [''],
-
-            },
-            /* Bengali
-         * -------
-         * 'va' and 'ba' are both rendered as ব.
-         */
-            bengali: {
-                vowels: 'অ আ ই ঈ উ ঊ ঋ ৠ ঌ ৡ  এ ঐ  ও ঔ'.split(' '),
-                vowel_marks: 'া ি ী ু ূ ৃ ৄ ৢ ৣ  ে ৈ  ো ৌ'.split(' '),
-                other_marks: 'ং ঃ ঁ'.split(' '),
-                virama: ['্'],
-                consonants: 'ক খ গ ঘ ঙ চ ছ জ ঝ ঞ ট ঠ ড ঢ ণ ত থ দ ধ ন প ফ ব ভ ম য র ল ব শ ষ স হ ळ ক্ষ জ্ঞ'.split(' '),
-                symbols: '০ ১ ২ ৩ ৪ ৫ ৬ ৭ ৮ ৯ ॐ ঽ । ॥'.split(' '),
-                other: '    ড ঢ  য '.split(' ')
-            },
-
-            newa: {
-                vowels: ['\u{11400}','\u{11401}',
-                    '\u{11402}','\u{11403}',
-                    '\u{11404}','\u{11405}',
-                    '\u{11406}','\u{11407}',
-                    '\u{11408}','\u{11409}',
[…]
