+++ matrix-editor/lib/sa.js
@@ -1,24 +1,0 @@
-(function () {
-var module = {
-    exports: null
-};
-
-module.exports = {
-    id : 'sa',
-    leftmin : 2,
-    rightmin : 2,
-    patterns : {
-        2 : "a1e1i1o1u1é1í1ï1ó1ú1ü1ā1ī1ū1ḷ1ḹ1ṛ1ṝ1",
-        3 : "2b_2c_2d_2g_2h_2j_2k_2l_2m_2n_2p_2r_2s_2t_2v_2y_2ñ_2ś_2ḍ_2ḥ12ḫ12ḷ_2ṁ12ṃ12ṅ_2ṇ_2ṭ_2ẖ1l̥1r̥1",
-        4 : "2bh_2ch_2dh_2gh_2jh_2kh_2m̐12ph_2rb_2rd_2rg_2rk_2rp_2rt_2rḍ_2rṭ_2th_2ḍh_2ṭh_a2i1a2u1a3ï1a3ü1l̥̄1r̥̄1"
-    },
-};
-
-var h = new window['Hypher'](module.exports);
-if(typeof module.exports.id === 'string') {
-    module.exports.id = [module.exports.id];
-}
-for (var i = 0; i < module.exports.id.length; i += 1) {
-  window['Hypher']['languages'][module.exports.id[i]] = h;
-}
-}());
