+++ matrix-editor/lib/hypher-nojquery.js
@@ -1,231 +1,0 @@
-(function () {
-
-var module = {
-    exports: null
-};
-/**
- * @constructor
- * @param {!{patterns: !Object, leftmin: !number, rightmin: !number}} language The language pattern file. Compatible with Hyphenator.js.
- */
-function Hypher(language) {
-    var exceptions = [],
-        i = 0;
-    /**
-     * @type {!Hypher.TrieNode}
-     */
-    this.trie = this.createTrie(language['patterns']);
-
-    /**
-     * @type {!number}
-     * @const
-     */
-    this.leftMin = language['leftmin'];
-
-    /**
-     * @type {!number}
-     * @const
-     */
-    this.rightMin = language['rightmin'];
-
-    /**
-     * @type {!Object.<string, !Array.<string>>}
-     */
-    this.exceptions = {};
-
-    if (language['exceptions']) {
-        exceptions = language['exceptions'].split(/,\s?/g);
-
-        for (; i < exceptions.length; i += 1) {
-            this.exceptions[exceptions[i].replace(/\u2027/g, '').toLowerCase()] = new RegExp('(' + exceptions[i].split('\u2027').join(')(') + ')', 'i');
-        }
-    }
-}
-
-/**
- * @typedef {{_points: !Array.<number>}}
- */
-Hypher.TrieNode;
-
-/**
- * Creates a trie from a language pattern.
- * @private
- * @param {!Object} patternObject An object with language patterns.
- * @return {!Hypher.TrieNode} An object trie.
- */
-Hypher.prototype.createTrie = function (patternObject) {
-    var size = 0,
-        i = 0,
-        c = 0,
-        p = 0,
-        chars = null,
-        points = null,
-        codePoint = null,
-        t = null,
-        tree = {
-            _points: []
-        },
-        patterns;
-
-    for (size in patternObject) {
-        if (patternObject.hasOwnProperty(size)) {
-            patterns = patternObject[size].match(new RegExp('.{1,' + (+size) + '}', 'g'));
-
-            for (i = 0; i < patterns.length; i += 1) {
-                chars = patterns[i].replace(/[0-9]/g, '').split('');
-                points = patterns[i].split(/\D/);
-                t = tree;
-
-                for (c = 0; c < chars.length; c += 1) {
-                    codePoint = chars[c].charCodeAt(0);
-
-                    if (!t[codePoint]) {
-                        t[codePoint] = {};
-                    }
-                    t = t[codePoint];
-                }
-
-                t._points = [];
-
-                for (p = 0; p < points.length; p += 1) {
-                    t._points[p] = points[p] || 0;
-                }
-            }
-        }
-    }
-    return tree;
-};
-
-/**
- * Hyphenates a text.
- *
- * @param {!string} str The text to hyphenate.
- * @return {!string} The same text with soft hyphens inserted in the right positions.
- */
-Hypher.prototype.hyphenateText = function (str, minLength) {
-    minLength = minLength || 4;
-
-    // Regexp("\b", "g") splits on word boundaries,
-    // compound separators and ZWNJ so we don't need
[…]
