+++ matrix-editor/src/README.md
@@ -1,0 +1,1 @@
+This matrix editor opens CSV and TEICorpus XML files.
