+++ matrix-editor/src/README.md
@@ -1,1 +1,0 @@
-This matrix editor opens CSV and TEICorpus XML files.
