### 2021-12-28 438495f
fixed hyphenation function
