+++ matrix-editor/src/script.js
@@ -1,44 +1,46 @@
 import { Sanscript } from './lib/sanscript.mjs';
-import { Normalizer } from './lib/normalize.mjs';
 import { CSV } from './lib/csv.mjs';
 import { Smits } from './lib/jsphylosvg-custom.mjs';
+
 import { lemmaXSLT, prettyXSLT, csvXSLT, matrixXSLT } from './lib/xsltsheet.mjs';
 import { treeXSLT, lgXSLT } from './lib/xsltsheet-tree.mjs';
-import { HypherSa } from './lib/sa.mjs';
 
 import { showSaveFilePicker } from 'https://cdn.jsdelivr.net/npm/native-file-system-adapter/mod.js';
 //import { showSaveFilePicker } from 'native-file-system-adapter';
 
+import { Normalizer } from './lib/normalize.mjs';
+import { to, changeScript } from './lib/transliterate.mjs';
+
 import Hypher from 'hypher';
+import { hyphenation_sa } from './lib/hypher-sa.mjs';
+
+'use strict';
 
 window.comboView = (function() {
-
-    'use strict';
     
-    var _filename;
-    var _xml;
-    const _teins = 'http://www.tei-c.org/ns/1.0';
-    //var _xml = document.createElementNS('http://www.w3.org/1999/xhtml','teiCorpus');
-    //var _xml = document.createElementNS('http://www.tei-c.org/ns/1.0','group');
-    //var _xml = document.createElementNS('http://www.w3.org/1999/xhtml','group');
-    //var _texts = [];
-    //var _editions = [];
-    var _treelist = new Map();
-    var _trees = [];
-    var _textboxes = [];
-    var _matrix;
-    //var _menus;
-    var _viewdiv;
-    var _descs;
-    var _maxlemma;
-    var _windows = [window];
-    var _dragged;
-    var _undo = [];
-    var _redo = [];
-    var _editing = false;
-    var _normalization = false;
-    const _scripts = ['iast','devanagari','telugu','grantha','malayalam'];
-    const hyphenator = new Hypher(HypherSa);
+    const _const = {
+        teins: 'http://www.tei-c.org/ns/1.0',
+        scripts: ['iast','devanagari','telugu','grantha','malayalam']
+    };
+
+    const _state = {
+        filename: null,
+        xml: null,
+        treelist: new Map(),
+        trees: [],
+        textboxes: [],
+        matrix: null,
+        viewdiv: null,
+        descs: null,
+        maxlemma: null,
+        windows: [window],
+        dragged: null,
+        undo: [],
+        redo: [],
+        editing: false
+    };
+
+    const Hyphenator = new Hypher(hyphenation_sa);
 
     /*** Pure functions ***/
 
@@ -56,23 +58,6 @@ window.comboView = (function() {
             .replace(/\s+/g,' ')
             .trim();
     };
-
-    /*
-    const findNextLemma = function(lemma) {
-        let nextlemma = lemma.nextElementSibling;
-        while(nextlemma) {
-            if(!lemma.classList.contains('invisible'))
-                return lemma.textContent;
-            nextlemma = nextlemma.nextElementSibling;
-        }
-        return '';
-    };
-    */
-    /*const findNextLemma2 = function(arr,n) {
-        for(let i=parseInt(n)+1;i<arr.length;i++)
-            if(arr[i] != '') return arr[i];
-    };
-    */
 
     const pickColour = function(fadeFraction, rgbColor1, rgbColor2, rgbColor3) {
     // from https://gist.github.com/gskema/2f56dc2e087894ffc756c11e6de1b5ed
@@ -105,331 +90,6 @@ window.comboView = (function() {
         return 'rgb(' + gradient.red + ',' + gradient.green + ',' + gradient.blue + ')';
     };
 
[…]
