+++ matrix-editor/src/script.js
@@ -2,14 +2,15 @@ import { Sanscript } from './lib/sanscript.mjs';
 import { CSV } from './lib/csv.mjs';
 import { Smits } from './lib/jsphylosvg-custom.mjs';
 
-import { lemmaXSLT, prettyXSLT, csvXSLT, matrixXSLT } from './lib/xsltsheet.mjs';
-import { treeXSLT, lgXSLT } from './lib/xsltsheet-tree.mjs';
+//import { lemmaXSLT, prettyXSLT, csvXSLT, matrixXSLT } from './lib/xsltsheet.mjs';
+//import { treeXSLT, lgXSLT } from './lib/xsltsheet-tree.mjs';
 
 import { showSaveFilePicker } from 'https://cdn.jsdelivr.net/npm/native-file-system-adapter/mod.js';
 //import { showSaveFilePicker } from 'native-file-system-adapter';
 
 import { Normalizer } from './lib/normalize.mjs';
 import { to, changeScript } from './lib/transliterate.mjs';
+import { xslt as _xslt } from './lib/xslt.mjs';
 
 import Hypher from 'hypher';
 import { hyphenation_sa } from './lib/hypher-sa.mjs';
@@ -41,23 +42,9 @@ window.comboView = (function() {
     };
 
     const Hyphenator = new Hypher(hyphenation_sa);
+    const xslt = new _xslt(_state, _const);
 
     /*** Pure functions ***/
-
-    const makeLgLemma = function(str) {
-        if(!str.startsWith('<lg')) return str;
-
-        const xslt_proc = makeXSLTProc(lgXSLT);
-        return XSLTransformString(str,xslt_proc).firstChild.data.slice(1);
-    };
-
-    const multiLemmaConcat = function(arr) {
-        return arr.map(lemma => {
-            return makeLgLemma(lemma);
-        }).join(' ')
-            .replace(/\s+/g,' ')
-            .trim();
-    };
 
     const pickColour = function(fadeFraction, rgbColor1, rgbColor2, rgbColor3) {
     // from https://gist.github.com/gskema/2f56dc2e087894ffc756c11e6de1b5ed
@@ -105,21 +92,6 @@ window.comboView = (function() {
             nums.add(lemma.dataset.n);
         }
         return nums;
-    };
-
-    const makeXSLTProc = function(sheet) {
-        var parser = new DOMParser();
-        const xslsheet = parser.parseFromString(sheet,'text/xml');
-        const xslt_proc = new XSLTProcessor();
-        xslt_proc.importStylesheet(xslsheet);
-        return xslt_proc;
-    };
-
-    const XSLTransformString = function(s,proc) {
-        const temp = _state.xml.createElementNS(_const.teins,'ab');
-        temp.innerHTML = s;
-        //temp.setAttribute('xmlns','http://www.w3.org/1999/xhtml');
-        return proc.transformToFragment(temp,document);
     };
 
     const removeBox = function() {
@@ -855,10 +827,9 @@ window.comboView = (function() {
         },
 
         xml: async function(doc) {
-            const xslt_proc = makeXSLTProc(prettyXSLT);
             const str = new XMLSerializer().serializeToString(
                 //  XSLTransformElement(_state.xml,xslt_proc)
-                xslt_proc.transformToDocument(doc)
+                xslt.sheets['xml'].transformToDocument(doc)
             );
             const file = new Blob([str], {type: 'text/xml;charset=utf-8'});
             const fileURL = find.basename() + '.xml';
@@ -871,11 +842,10 @@ window.comboView = (function() {
         },
 
         csv: async function(doc) {
-            const xslt_proc = makeXSLTProc(csvXSLT);
             //const str = new XMLSerializer().serializeToString(XSLTransformElement(doc,xslt_proc));
[…]
