+++ matrix-editor/src/script.js
@@ -9,10 +9,11 @@ import { showSaveFilePicker } from 'https://cdn.jsdelivr.net/npm/native-file-sys
 //import { showSaveFilePicker } from 'native-file-system-adapter';
 
 import { Normalizer } from './lib/normalize.mjs';
-import { to, changeScript } from './lib/transliterate.mjs';
-import { xslt as _xslt } from './lib/xslt.mjs';
+import { changeScript } from './lib/transliterate.mjs';
+import { xslt as _Xslt } from './lib/xslt.mjs';
+import { find as _Find } from './lib/find.mjs';
 
-import Hypher from 'hypher';
+import _Hypher from 'hypher';
 import { hyphenation_sa } from './lib/hypher-sa.mjs';
 
 'use strict';
@@ -41,8 +42,9 @@ window.comboView = (function() {
         editing: false
     };
 
-    const Hyphenator = new Hypher(hyphenation_sa);
-    const xslt = new _xslt(_state, _const);
+    const Hypher = new _Hypher(hyphenation_sa);
+    const Xslt = new _Xslt(_state, _const);
+    const Find = new _Find(_state);
 
     /*** Pure functions ***/
 
@@ -201,7 +203,7 @@ window.comboView = (function() {
     };
 
     const touchUpText = function(str) {
-        return Hyphenator.hyphenateText(
+        return Hypher.hyphenateText(
             str
                 .replace(/ \|/g,'\u00a0|')
                 .replace(/\| (?=\d)/g,'|\u00a0')
@@ -246,7 +248,7 @@ window.comboView = (function() {
             if(!nums || nums.size === 0) return;
             multi.unHighlightAll();
         
-            const [low,high] = find.lowhigh(nums);
+            const [low,high] = Find.lowhigh(nums);
             if(high !== undefined) {
                 for(let n=low;n<=high;n++) multi.highlightLemma(n,true);
             }
@@ -261,7 +263,7 @@ window.comboView = (function() {
 
         highlightAll: function() {
             if(check.anyhighlit()) multi.unHighlightAll();
-            else multi.highlightRange(new Set([0,find.maxlemma()]));
+            else multi.highlightRange(new Set([0,Find.maxlemma()]));
         },
 
         unHighlightAll: function() {
@@ -493,13 +495,13 @@ window.comboView = (function() {
         }
         //_texts = new Map(csvarr);
         //_state.maxlemma = _texts.get(_texts.keys().next().value).text.length;
-        //_state.maxlemma = find.firsttext().lastElementChild.getAttribute('n');
-        _state.maxlemma = find.maxlemma();
+        //_state.maxlemma = Find.firsttext().lastElementChild.getAttribute('n');
+        _state.maxlemma = Find.maxlemma();
 
         //var mss = Array.from(_texts.keys());
         //mss.sort();
         var msshtml = '';
-        const mss = [...find.teis()].map(el => el.getAttribute('n'));
+        const mss = [...Find.teis()].map(el => el.getAttribute('n'));
         for(const ms of mss)
         //    msshtml += `<li data-name="${ms}">${_texts.get(ms).desc}</li>`;
             msshtml += `<li data-name="${ms}">${ms}</li>`;
@@ -613,7 +615,7 @@ window.comboView = (function() {
         }
         _state.xml.normalize();
 
-        _state.maxlemma = find.maxlemma();
+        _state.maxlemma = Find.maxlemma();
         
         //if(fs.length > 0) csvLoadAdditional(fs);
 
@@ -631,7 +633,7 @@ window.comboView = (function() {
             treeXMLLoad(nexml,tree.closest('stemma').id,false);
         }
 
-        _state.maxlemma = find.maxlemma();
+        _state.maxlemma = Find.maxlemma();
[…]
