+++ matrix-editor/src/script.js
@@ -12,6 +12,7 @@ import { Normalizer } from './lib/normalize.mjs';
 import { changeScript } from './lib/transliterate.mjs';
 import { xslt as _Xslt } from './lib/xslt.mjs';
 import { find as _Find } from './lib/find.mjs';
+import { workerFunc } from './lib/worker.mjs';
 
 import _Hypher from 'hypher';
 import { hyphenation_sa } from './lib/hypher-sa.mjs';
@@ -2062,7 +2063,7 @@ const fullTreeClick = function(e) {
             _state.matrix.boxdiv.querySelector('tbody').appendChild(tr);
             th.scrollIntoView();
 
-            const workerblob = new Blob(['('+worker.fitch.toString()+')()'],{type: 'application/javascript'});
+            const workerblob = new Blob(['('+workerFunc.toString()+')()'],{type: 'application/javascript'});
             const fitchWorker = new Worker(window.URL.createObjectURL(workerblob));
             const normalized = check.normalizedView();
             const serialreadings = Find.serializedtexts(tree.nexml,normalized);
@@ -4102,124 +4103,6 @@ const fullTreeClick = function(e) {
             this.boxdiv.append(scroller);
         }
     }
-
-    const worker = {
-        fitch: function() {
-            var mss,levels,num,id;
-
-            const find = {
-                setIntersection: function(...sets) {
-                    const setA = sets[0];
-                    return new Set(
-                        [...setA].filter(el => {
-                            for(let n=1;n<sets.length;n++) {
-                                if(!sets[n].has(el))
-                                    return false;
-                            }
-                            return true;
-                        })
-                    );
-                },
-
-                setUnion: function(...sets) {
-                    return new Set(
-                        sets.reduce((acc, cur) => {
-                            acc = [...acc,...cur];
-                            return acc;
-                        },[])
-                    );
-                },
-            };
-
-            const fitch1 = function(n,target) {
-                const firstpass = new Map();
-                for(const taxon of levels[0]) {
-                    const val = mss.get(taxon);
-                    //const reading = mss.get(taxon)[n];
-                    //const reading = val? val[n] : '';
-                    const reading = val ? [val[n]] : [];
-                    firstpass.set(taxon,new Set(reading));
-                }
-                for(let m=1;m<levels.length;m++) { // start at 1 (after taxa)
-                    for(const [node,children] of levels[m]) {
-                        const readings = children.map(node => firstpass.get(node));
-                        const intersection = find.setIntersection(...readings);
-                    
-                        // shortcut, because we only care about one node
-                        if(node === target && intersection.size === 1)
-                            return [...intersection][0];
-
-                        const result = intersection.size > 0 ?
-                            intersection :
-                            find.setUnion(...readings);
-                        firstpass.set(node,result);
-
-                    }
-                }
-                return firstpass;
-
-            };
-
-            const fitch2 = function(firstpass,target) {
-                const taxa = levels[0];
-                const secondpass = new Map();
-                
-                /*
-                for(const [node,children] of levels[levels.length-1]) {
-                    secondpass.set(node,firstpass.get(node));
-                }
-                */
-                for(const level of levels[levels.length-1])
-                    secondpass.set(level[0],firstpass.get(level[0]));
-
-                for(let n=levels.length-1;n>1;n--) {
-                    for(const [node,children] of levels[n]) {
-                        const ancestral = secondpass.get(node);
-                        for(const child of children) {
-                            if(taxa.indexOf(child) !== -1)
-                                continue;
-                            const childreading = firstpass.get(child);
-                            const intersection = find.setIntersection(ancestral,childreading);
-                            const result = intersection.size > 0 ?
-                                intersection :
-                                childreading;
-                        
[…]
