+++ matrix-editor/src/script.js
@@ -2,9 +2,6 @@ import { Sanscript } from './lib/sanscript.mjs';
 import { CSV } from './lib/csv.mjs';
 import { Smits } from './lib/jsphylosvg-custom.mjs';
 
-//import { lemmaXSLT, prettyXSLT, csvXSLT, matrixXSLT } from './lib/xsltsheet.mjs';
-//import { treeXSLT, lgXSLT } from './lib/xsltsheet-tree.mjs';
-
 import { showSaveFilePicker } from 'https://cdn.jsdelivr.net/npm/native-file-system-adapter/mod.js';
 //import { showSaveFilePicker } from 'native-file-system-adapter';
 
@@ -12,6 +9,7 @@ import { Normalizer } from './lib/normalize.mjs';
 import { changeScript } from './lib/transliterate.mjs';
 import { xslt as _Xslt } from './lib/xslt.mjs';
 import { find as _Find } from './lib/find.mjs';
+import { check as _Check } from './lib/check.mjs';
 import { workerFunc } from './lib/worker.mjs';
 
 import _Hypher from 'hypher';
@@ -46,6 +44,7 @@ window.comboView = (function() {
     const Hypher = new _Hypher(hyphenation_sa);
     const Xslt = new _Xslt(_state, _const);
     const Find = new _Find(_state);
+    const Check = new _Check(_state, Find);
 
     const removeBox = function() {
         const box = document.getElementById('tooltip');
@@ -213,7 +212,7 @@ window.comboView = (function() {
         },
 
         highlightAll: function() {
-            if(check.anyhighlit()) multi.unHighlightAll();
+            if(Check.anyhighlit()) multi.unHighlightAll();
             else multi.highlightRange(new Set([0,Find.maxlemma()]));
         },
 
@@ -384,161 +383,7 @@ window.comboView = (function() {
             if(show) newBox.tree(stemmaid,id);
         }
     };
-    /*
-    const csvLoad = function(f,fs,e) {
-        _state.filename = f.name;
-        const ext = _state.filename.split('.').pop();
-        var csvarr = [];
-        if(ext === 'csv') {
-            _state.xml = document.implementation.createDocument(_const.teins,'',null);
-            const teicorpus = _state.xml.createElementNS(_const.teins,'teiCorpus');
-            const teiheader = _state.xml.createElementNS(_const.teins,'teiHeader');
-            teicorpus.appendChild(teiheader);
-            _state.xml.appendChild(teicorpus);
-            const csvstr = e.target.result;
-            csvarr = csvstr.trim().split(/\n+/)
-                .map(s => s.replace(/""/g,'').split(','))
-                .map(a => {
-                    const name = a.shift();
-                    const tei = _state.xml.createElementNS(_const.teins,'TEI');
-                    //tei.setAttribute('xmlns','http://www.tei-c.org/ns/1.0');
-                    tei.setAttribute('n',name);
-                    const text = _state.xml.createElementNS(_const.teins,'text');
-                    //text.setAttribute('xmlns','http://www.tei-c.org/ns/1.0');
-                    //const root = document.createElementNS('http://www.w3.org/1999/xhtml','text');
-                    //root.setAttribute('n',name);
-                    for(let n=0;n<a.length;n++) {
-                        const word = a[n];
-                        //const newEl = document.createElementNS('http://www.tei-c.org/ns/1.0','w');
-                        //newEl.setAttribute('xmlns','http://www.tei-c.org/ns/1.0');
-                        const newEl = _state.xml.createElementNS(_const.teins,'w');
-                        //newEl.setAttribute('xmlns','http://www.w3.org/1999/xhtml');
-                        if(word !== '')
-                            newEl.appendChild(document.createTextNode(word));
-                        newEl.setAttribute('n',n);
[…]
