### 2022-01-02 6ea691b
cleaned up webworker
