+++ matrix-editor/script.js
@@ -101,5 +101,5 @@ END;
         <button type="submit">Add to matrix</button>
       </div>
     </form>
-</div>`);r.appendChild(a),document.body.appendChild(r);let h=function(v){v.preventDefault();let B=r.querySelector("input"),k=B.value?B.value:B.placeholder;q.doReconstruction(n,l,k),document.body.removeChild(r)};r.querySelector("button").addEventListener("click",h),r.addEventListener("click",fe.blackoutClick)},thKeyDown:function(s){s.key==="Enter"&&q.finishNewRow(s)},finishNewRow:function(s){let l=s.target,n=l.textContent;l.closest("tr").dataset.n=n;let t=ve.tei(n);f.xml.documentElement.appendChild(t),f.editing=!1,l.contentEditable=!1,l.removeEventListener("blur",q.finishNewRow),l.removeEventListener("keydown",q.thKeyDown),q.doStack([q.doDeleteRow,[n]],"do")},finishEditCell:function(s,l=!1){let n=s.target;f.editing=!1,n.contentEditable="false",n.removeEventListener("blur",q.finishEditCell),n.removeEventListener("keydown",q.cellKeyDown),n.blur(),de.deselect();let t=n.textContent;if(t===""){let g=n.querySelector("br");g&&g.remove()}if(l||t===n.dataset.oldContent)return;n.hasOwnProperty("IAST")||(n.IAST=n.cloneNode(!0)),n.IAST.textContent=t;let r=parseInt(n.dataset.n),a=n.closest("tr"),h=a.dataset.n;if(q.xmlChangeCell(r,h,t),a.dataset.hasOwnProperty("treename")&&!n.dataset.hasOwnProperty("emended")){let g=q.doEmend(r,h,"multido"),v=[];v.push([q.doChangeCell,[r,h,n.dataset.oldContent]]),v.push(g),q.doStack([q.doMulti,[v]],"do")}else q.doStack([q.doChangeCell,[r,h,n.dataset.oldContent]],"do");delete n.dataset.oldContent,q.refresh(),ye.updateAllHeaders(!0),n.classList.add("highlitcell")},finishMoveRow:function(s){let l=s.target.nodeType===1?s.target.closest("tr"):s.target.parentElement.closest("tr");if(l&&q.doMoveRow(f.dragged,l,"do"),s.type!=="drop"){for(let n of p.trs())n.classList.remove("moveinprogress");f.matrix.boxdiv.querySelector("table").removeEventListener("mousedown",q.finishMoveRow)}else for(let n of p.trs())n.classList.remove("dragenter");f.dragged.classList.remove("dragging"),f.dragged=null},doReconstruction:function(s,l,n){let t=ve.row(n,"pending");t.dataset.treename=s.name,t.dataset.nodename=l;let r=t.querySelector("th"),a=document.createElement("div");a.id="spinner",r.prepend(a);let h=ve.tei(n);h.setAttribute("type","reconstructed"),h.setAttribute("corresp",s.name),h.setAttribute("select",`#${l}`);let g=[...p.tds(!1,t)],v=[...p.words(!1,h)];f.matrix.boxdiv.querySelector("tbody").appendChild(t),r.scrollIntoView();let B=new Worker("./worker.js",{type:"module"}),k=N.normalizedView(),D=p.serializedtexts(s.nexml,k),V=new Map(D.map(Q=>[Q[0],Q[1][0]]));B.postMessage({readings:V,levels:s.levels,num:0,id:l}),B.onmessage=function(Q){let te=Q.data.n,ue=Q.data.result;if
[…]
