### 2022-01-02 cdeb378
fixed matrix click being called twice
