### 2022-01-02 b8e37d6
Fixed double text click call
