+++ matrix-editor/script.js
@@ -91,7 +91,7 @@ END;
         <button type="submit">Export</button>
       </div>
     </form>
-</div>`),l=s.getElementById("option__basetext"),n=[...p.trs()].map(t=>t.dataset.n);for(let t of n){let r=document.createElement("option");r.value=t,r.append(t),l.appendChild(r)}return s},showOptions:function(s,l){let n=l(),t=document.createElement("div");t.id="blackout",t.appendChild(n),document.body.appendChild(t);let r=function(h){h.preventDefault();var g=[];let v=document.getElementById("exportform").elements;for(let k of v)k.type==="checkbox"?g.push([k.name,k.checked]):g.push([k.name,k.value]);let B=fe.processOptions(new Map(g));s(B),document.body.removeChild(t)};return t.querySelector("button").addEventListener("click",r),t.addEventListener("click",fe.blackoutClick),t},blackoutClick:function(s){if(!s.target.closest(".popup")){let n=document.querySelector("#blackout");n.parentNode.removeChild(n)}}},de={selectAll(s){let l=document.createRange();l.selectNodeContents(s);let n=window.getSelection();n.removeAllRanges(),n.addRange(l)},deselect(){let s=window.getSelection();s.removeAllRanges?s.removeAllRanges():s.empty&&s.empty()},treeMouseover(s){let l=s.target.classList.contains("tree-lemma")?s.target:s.target.closest(".tree-lemma");l&&(L.highlightTreeLemma(l.dataset.id),l.addEventListener("mouseout",L.unhighlightTrees));let n=s.target.dataset.reconstructed;if(!n)return;let t=document.createElement("div");t.id="tooltip",t.style.top=s.pageY+"px",t.style.left=s.pageX+"px",t.style.opacity=0,t.style.transition="opacity 0.2s ease-in",f.viewdiv.parentElement.appendChild(t);let r=document.createElement("div");if(r.appendChild(document.createTextNode(n)),this.script!==0?t.appendChild(Ot(r,x.scripts[this.script])):t.appendChild(r),s.target.classList.contains("reconstructed")){let a=s.target.parentNode.querySelector("span.tree-lemma");if(a&&a.dataset.hasOwnProperty("emended")){let h=document.createElement("div");h.classList.add("emphasis"),h.appendChild(document.createTextNode(a.textContent)),this.script!==0?t.prepend(Ot(h,x.scripts[this.script])):t.prepend(h)}}window.getComputedStyle(t).opacity,t.style.opacity=1,s.target.addEventListener("mouseout",u)},treeClick(s){s.target.classList.contains("witness")?b.text(s.target.dataset.key):s.target.classList.contains("reconstructed")?b.text(s.target.dataset.label):s.target.classList.contains("internal")&&q.startReconstruction(s),u()},keyDown(s){if(!f.editing&&s.key.substring(0,5)==="Arrow")de.cycleVariant(s);else if(s.ctrlKey||s.metaKey)s.key==="Z"?q.redo():s.key==="z"&&q.undo();else if(!f.editing&&f.matrix&&!f.matrix.closed&s.key==="Enter"){let l=p.highlitcell();l&&(s.preventDefault(),q.startEditCell(l))}},cycleVariant(s){let l=p.highlitcell()||f.viewdiv.querySelector("td.highlit")||f.viewdiv.querySelector('td[data-n="0"]');switch(s.key){case"ArrowRight":{if(!f.matrix.closed&&l)
[…]
