+++ matrix-editor/src/script.js
@@ -2,14 +2,12 @@ import { Sanscript } from './lib/sanscript.mjs';
 import { CSV } from './lib/csv.mjs';
 import { Smits } from './lib/jsphylosvg-custom.mjs';
 
-import { showSaveFilePicker } from 'https://cdn.jsdelivr.net/npm/native-file-system-adapter/mod.js';
-//import { showSaveFilePicker } from 'native-file-system-adapter';
-
 import { Normalizer } from './lib/normalize.mjs';
 import { changeScript } from './lib/transliterate.mjs';
 import { xslt as _Xslt } from './lib/xslt.mjs';
 import { Utils as _Utils } from './lib/utils.mjs';
 import { Fitch as _Fitch } from './lib/fitch.mjs';
+import { Exporter as _Exporter } from './lib/export.mjs';
 
 import _Hypher from 'hypher';
 import { hyphenation_sa } from './lib/hypher-sa.mjs';
@@ -18,12 +16,9 @@ import { hyphenation_sa } from './lib/hypher-sa.mjs';
 
 window.comboView = (function() {
     
-    const _const = {
-        teins: 'http://www.tei-c.org/ns/1.0',
-        scripts: ['iast','devanagari','telugu','grantha','malayalam']
-    };
-
     const _state = {
+        teins: 'http://www.tei-c.org/ns/1.0',
+        scripts: ['iast','devanagari','telugu','grantha','malayalam'],
         filename: null,
         xml: null,
         treelist: new Map(),
@@ -41,10 +36,12 @@ window.comboView = (function() {
     };
 
     const Hypher = new _Hypher(hyphenation_sa);
-    const Xslt = new _Xslt(_state, _const);
+    const Xslt = new _Xslt(_state);
     const Utils = new _Utils(_state);
     const Find = Utils.find;
     const Check = Utils.check;
+    const Make = Utils.make;
+    const Exporter = new _Exporter(Utils, Xslt); // change to dynamic import?
 
     const removeBox = function() {
         const box = document.getElementById('tooltip');
@@ -336,14 +333,14 @@ window.comboView = (function() {
         const nexml = parser.parseFromString(treestr,'text/xml');
         const xenoData = _state.xml.querySelector('teiHeader > xenoData') || (function() {
             const header = _state.xml.querySelector('teiHeader') || (function() {
-                const h = _state.xml.createElementNS(_const.teins,'teiHeader');
+                const h = _state.xml.createElementNS(_state.teins,'teiHeader');
                 _state.xml.documentElement.appendChild(h);
                 return h;})();
-            const newel = _state.xml.createElementNS(_const.teins,'xenoData');
+            const newel = _state.xml.createElementNS(_state.teins,'xenoData');
             header.appendChild(newel);
             return newel;
         })();
-        const stemmael = _state.xml.createElementNS(_const.teins,'stemma');
+        const stemmael = _state.xml.createElementNS(_state.teins,'stemma');
         stemmael.setAttribute('format','nexml');
         stemmael.id = 'stemma' + [...xenoData.querySelectorAll('stemma')].length;
         stemmael.appendChild(nexml.firstChild.cloneNode(true));
@@ -386,21 +383,21 @@ window.comboView = (function() {
 
     const csvLoad = function(f,fs,e) {
         const csvarr = CSV.parse(e.target.result,{delimiter: ','});
-        _state.xml = document.implementation.createDocument(_const.teins,'',null);
-        const teicorpus = _state.xml.createElementNS(_const.teins,'teiCorpus');
-        const teiheader = _state.xml.createElementNS(_const.teins,'teiHeader');
+        _state.xml = document.implementation.createDocument(_state.teins,'',null);
[…]
