+++ matrix-editor/src/script.js
@@ -1,4 +1,3 @@
-import { Sanscript } from './lib/sanscript.mjs';
 import { CSV } from './lib/csv.mjs';
 import { Smits } from './lib/jsphylosvg-custom.mjs';
 
@@ -46,44 +45,6 @@ window.comboView = (function() {
     const removeBox = function() {
         const box = document.getElementById('tooltip');
         if(box) box.remove();
-    };
-
-    const fillSelector = function() {
-        const menudiv = document.getElementById('left_menu'); 
-        menudiv.innerHTML =
- `<div id="matrixmenu" class="menubox">
-    <div class="heading">Matrix</div>
-      <ul>
-        <li>select CSV/XML file(s): <input type="file" id="file" name="file" accept=".xml,.csv" multiple/></li>
-      </ul>
-</div>
-<div style="display: none" id="mssmenu" class="menubox">
-    <div class="heading">Texts</div>
-    <ul class="ms">
-    </ul>
-</div>
-<div style="display: none" id="treemenu" class="menubox">
-    <div class="heading">Trees</div>
-        <ul class="tree">
-            <li>select NeXML file: <input type="file" id="treefile" name="treefile" accept=".xml"/></li>
-        </ul>
-    </div>
-`;
-        /*
-    var mss = Array.from(_texts.keys());
-    mss.sort();
-    var msshtml = '';
-    for(const ms of mss)
-        msshtml += `<li data-name="${ms}">${_texts.get(ms).desc}</li>`;
-    menudiv.querySelector('.ms').innerHTML = msshtml;
-*/
-        const menu = document.getElementById('menu');
-        menu.addEventListener('mouseover', events.menuMouseover);
-        menu.addEventListener('mouseout', events.menuMouseout);
-        menu.addEventListener('click',events.menuClick);
-        // }
-        menu.querySelector('#file').addEventListener('change',fileSelect.bind(null,csvOrXml),false);
-        menu.querySelector('#treefile').addEventListener('change',fileSelect.bind(null,treeFileLoad),false);
     };
 
     const newBox = {
@@ -333,14 +294,14 @@ window.comboView = (function() {
         const nexml = parser.parseFromString(treestr,'text/xml');
         const xenoData = _state.xml.querySelector('teiHeader > xenoData') || (function() {
             const header = _state.xml.querySelector('teiHeader') || (function() {
-                const h = _state.xml.createElementNS(_state.teins,'teiHeader');
+                const h = Make.xmlel('teiHeader');
                 _state.xml.documentElement.appendChild(h);
                 return h;})();
-            const newel = _state.xml.createElementNS(_state.teins,'xenoData');
+            const newel = Make.xmlel('xenoData');
             header.appendChild(newel);
             return newel;
         })();
-        const stemmael = _state.xml.createElementNS(_state.teins,'stemma');
+        const stemmael = Make.xmlel('stemma');
         stemmael.setAttribute('format','nexml');
         stemmael.id = 'stemma' + [...xenoData.querySelectorAll('stemma')].length;
         stemmael.appendChild(nexml.firstChild.cloneNode(true));
@@ -384,20 +345,20 @@ window.comboView = (function() {
     const csvLoad = function(f,fs,e) {
         const csvarr = CSV.parse(e.target.result,{delimiter: ','});
         _state.xml = document.implementation.createDocument(_state.teins,'',null);
-        const teicorpus = _state.xml.createElementNS(_state.teins,'teiCorpus');
-        const teiheader = _state.xml.createElementNS(_state.teins,'teiHeader');
+        const teicorpus = Make.xmlel('teiCorpus');
+        const teiheader = Make.xmlel('teiHeader');
         teicorpus.appendChild(teiheader);
[…]
