### 2022-01-09 763c060
quick fix for column removal
