+++ matrix-editor/src/script.js
@@ -503,7 +503,7 @@ window.comboView = (function() {
             },
             {text: 'Delete column',
                 greyout: Check.anyhighlit,
-                func: edit.startRemoveCol.bind(null,false)
+                func: edit.removeCol.start.bind(null,false)
             },
             {text: 'Delete empty columns',
                 func: edit.startRemoveEmpty
@@ -530,7 +530,7 @@ window.comboView = (function() {
             },
             {text: 'Insert column',
                 greyout: Check.anyhighlit,
-                func: edit.startInsertCol
+                func: edit.insertCol.start
             },
             {text: 'Mark insignificant',
                 checkbox: Check.checkbox.bind(null,'insignificant'),
@@ -1028,7 +1028,7 @@ const fullTreeClick = function(e) {
                             func: edit.group.start.bind(null,false)
                         },
                         {text: 'delete columns',
-                            func: edit.startRemoveCol.bind(null,nums)
+                            func: edit.removeCol.start.bind(null,nums)
                         },
                         /*                {text: 'insignificant',
                      cond: Check.checkbox.bind(null,'insignificant',nums),
@@ -1044,10 +1044,10 @@ const fullTreeClick = function(e) {
                             func: edit.startEditCell.bind(null,td)
                         },
                         {text: 'delete column',
-                            func: edit.startRemoveCol.bind(null,nums)
+                            func: edit.removeCol.start.bind(null,nums)
                         },
                         {text: 'insert column',
-                            func: edit.startInsertCol
+                            func: edit.insertCol.start
                         },
                         /*                {text: 'insignificant',
                      cond: Check.checkbox.bind(null,'insignificant',nums),
@@ -1165,7 +1165,6 @@ const fullTreeClick = function(e) {
         },
         
         ungroup: {
-
             start: function(nums) {
                 const clgroups = Find.clauses(nums);
                 const args = [...clgroups].map(s => [edit.ungroup.go,[s]]);
@@ -1185,45 +1184,112 @@ const fullTreeClick = function(e) {
         
         },
         
-        startInsertCol: function() {
-            const insertafter = Math.max([...Find.highlit()]) || _state.maxlemma;
-            edit.doInsertCol(insertafter);
+        insertCol: {
+            start: function() {
+                const insertafter = Math.max([...Find.highlit()]) || _state.maxlemma;
+                edit.insertCol.go(insertafter);
+            },
+            go: function(aftern,doing = 'do') {
+                const insert = function(rowfunc,cellfunc,newfunc,aftern) {
+                    const rows = rowfunc();
+                    for(const row of rows) {
+                        const cell = cellfunc(aftern,row);
+                        const newcell = newfunc(aftern + 1);
+                        cell.insertAdjacentElement('afterend',newcell);
+                    }
+                };
+                insert(Find.trs,Find.firsttd,Make.emptycell,aftern);
+                insert(Find.texts,Find.firstword,Make.emptyword,aftern);
+                edit.renumber(aftern);
+
+                //view.renormalize(aftern,aftern+2);
+                edit.refresh();
+                edit.restyleGroups([aftern,aftern+1],true);
+                view.updateAllHeaders();
+                
+                const arg = [aftern+1];
+                if(doing === 'multido')
+                    return [edit.removeCol.go,[arg]];
+                else
+                    edit.doStack([edit.removeCol.go,[arg]],doing);
+            }
         },
+        
+        removeCol: {
+            start: function(nums,/*e*/) {
+                const numss = nums === false ?
+                    Find.highlit() :
[…]
