+++ matrix-editor/src/script.js
@@ -502,7 +502,9 @@ window.comboView = (function() {
                 func: edit.startRenormalize.bind(null,false)
             },
             {text: 'Delete column',
+                alt: 'Delete columns',
                 greyout: Check.anyhighlit,
+                toggle: Check.manyhighlit,
                 func: edit.removeCol.start.bind(null,false)
             },
             {text: 'Delete empty columns',
@@ -1187,31 +1189,68 @@ const fullTreeClick = function(e) {
         insertCol: {
             start: function() {
                 const insertafter = Math.max([...Find.highlit()]) || _state.maxlemma;
-                edit.insertCol.go(insertafter);
+                edit.insertCol.go({mode: 'append', start: insertafter});
             },
-            go: function(aftern,doing = 'do') {
-                const insert = function(rowfunc,cellfunc,newfunc,aftern) {
+            go: function(opts,doing = 'do') {
+                // opts = { mode: 'append' || 'insert', start, xml, html }
+                const insert = function(rowfunc,cellfunc,els,newfunc) {
                     const rows = rowfunc();
-                    for(const row of rows) {
-                        const cell = cellfunc(aftern,row);
-                        const newcell = newfunc(aftern + 1);
-                        cell.insertAdjacentElement('afterend',newcell);
+                    for(var n=0;n<rows.length;n++) {
+                        const anchor = cellfunc(opts.start,rows[n]);
+                        if(opts.mode === 'append') {
+                            if(!els) {
+                                const newcell = newfunc(opts.start + 1);
+                                anchor.insertAdjacentElement('afterend',newcell);
+                            }
+                            else {
+                                for(const cell of els[n].reverse())
+                                    anchor.insertAdjacentElement('afterend',cell);
+                            }
+                        }
+                        else {
+                            if(!els) {
+                                const newcell = newfunc(opts.start);
+                                anchor.insertAdjacentElement('beforebegin',newcell);
+                            }
+                            else {
+                                for(const cell of els[n])
+                                    anchor.insertAdjacentElement('beforebegin',cell);
+                            }
+                        }
                     }
+                    return els ? els[0].length : 1;
                 };
-                insert(Find.trs,Find.firsttd,Make.emptycell,aftern);
-                insert(Find.texts,Find.firstword,Make.emptyword,aftern);
-                edit.renumber(aftern);
 
-                //view.renormalize(aftern,aftern+2);
+                multi.unHighlightAll();
+
+                var len;
+                if(opts.hasOwnProperty('xml')) {
+                    insert(...opts.xml);
+                    len = insert(...opts.html);
+                }
+                else {
+                    insert(Find.trs,Find.firsttd,null,Make.emptycell);
+                    insert(Find.texts,Find.firstword,null,Make.emptyword);
+                    len = 1;
+                }
+
+                const nums = opts.mode === 'append' ? 
+                    [...Array(len).keys()].map(n => opts.start + n + 1) : 
+                    [...Array(len).keys()].map(n => opts.start + n);
+
+                edit.renumber(opts.start);
+
+                multi.highlightRange(new Set(nums));
+
                 edit.refresh();
-                edit.restyleGroups([aftern,aftern+1],true);
+                edit.reIAST(nums);
+                edit.restyleGroups(nums,true);
                 view.updateAllHeaders();
-                
-                const arg = [aftern+1];
+
                 if(doing === 'multido')
-                    return [edit.removeCol.go,[arg]];
+                    return [edit.removeCol.go,[nums]];
                 else
[…]
