### 2022-01-09 e49474e
cleaned up menus
