+++ matrix-editor/src/script.js
@@ -497,8 +497,20 @@ window.comboView = (function() {
                 toggle: Check.anyhighlit,
                 func: multi.highlightAll
             },
-            {text: '(Re)normalize columns',
+            {text: 'Group all words',
+                func: edit.doGroupWords
+            },
+            {text: 'Delete empty columns',
+                func: edit.startRemoveEmpty
+            }
+        ]);
+
+        const columnbox = new menuBox('Column');
+        columnbox.populate([
+            {text: '(Re)normalize column',
+                alt: '(Re)normalize columns',
                 greyout: Check.anyhighlit,
+                toggle: Check.manyhighlit,
                 func: edit.startRenormalize.bind(null,false)
             },
             {text: 'Delete column',
@@ -506,9 +518,6 @@ window.comboView = (function() {
                 greyout: Check.anyhighlit,
                 toggle: Check.manyhighlit,
                 func: edit.removeCol.start.bind(null,false)
-            },
-            {text: 'Delete empty columns',
-                func: edit.startRemoveEmpty
             },
             {text: 'Merge columns',
                 greyout: Check.manyhighlit,
@@ -519,16 +528,6 @@ window.comboView = (function() {
                 greyout: Check.oneGrouped,
                 toggle: Check.grouped,
                 func: edit.group.start.bind(null,false),
-            },
-            {text: 'Group words',
-                func: edit.doGroupWords
-            },
-            {text: 'Edit cell',
-                greyout: Check.highlitcell,
-                func: edit.startEditCell.bind(null,false),
-            },
-            {text: 'Insert row',
-                func: edit.startNewRow,
             },
             {text: 'Insert column',
                 greyout: Check.anyhighlit,
@@ -543,9 +542,24 @@ window.comboView = (function() {
                 checkbox: Check.checkbox.bind(null,'binary'),
                 greyout: Check.anyhighlit,
                 func: edit.startMarkAs.bind(null,'binary',false)
-            },
+            }
         ]);
 
+        const rowbox = new menuBox('Row');
+        rowbox.populate([
+            {text: 'Insert row',
+                func: edit.startNewRow,
+            }
+        ]);
+
+        const cellbox = new menuBox('Cell');
+        cellbox.populate([
+            {text: 'Edit cell',
+                greyout: Check.highlitcell,
+                func: edit.startEditCell.bind(null,false),
+            }
+        ]);
+        
         const viewbox = new menuBox('View');
         viewbox.populate([
             {text: 'Header',
@@ -562,6 +576,9 @@ window.comboView = (function() {
         const left_menu = document.getElementById('left_menu');
         left_menu.appendChild(viewbox.box);
         left_menu.appendChild(editbox.box);
+        left_menu.appendChild(columnbox.box);
+        left_menu.appendChild(rowbox.box);
+        left_menu.appendChild(cellbox.box);
         left_menu.appendChild(expbox.box);
    
         const views = document.getElementById('views');
@@ -1279,7 +1296,7 @@ const fullTreeClick = function(e) {
             go: function(nums,doing = 'do') {
                 const remove = function(rowfunc,cellfunc) {
                     const rows = rowfunc();
-                    var rowsclone = [];
+                    const rowsclone = [];
                     for(const row of rows) {
                         const arr = [...nums].map(n => {
                             const cell = cellfunc(n,row);
@@ -1293,36 +1310,32 @@ const fullTreeClick = function(e) {
                     }
                     return [rowfunc,cellfunc,rowsclone];
                 };
-                //const oldhtml = remove(document,'.matrix tr','td','data-n',nums);
-                //const oldxml = remove(_state.xml,'text','w','n',nums);
-                const oldmax = Find.maxlemma();
[…]
