### 2022-01-12 #4022 @galessiorob
Thank you so much for your thorough and speedy review mid holidays @kinow 🎉 !

I agree with the notes and suggestions you propose in your [PR](https://github.com/chchch/sanskrit-alignment/pull/5), @chchch mind reviewing the recommendations and considering adding tests for the `matrix-editor`?

I'll review the paper itself again shortly.
