### 2022-01-16 #4022 @chchch
Hello,

I've started adding tests to the main branch. Should I be keeping the joss-paper branch synced with the main branch?
