### 2022-02-09 chchch/sanskrit-alignment#7 @xiaohk
Inconsistent license?
Hello, I am a JOSS reviewer for helayo. Using sequence alignment on ext data looks very cool! The UI is very neat too.

I am a bit confused about the license for this project. The [document](https://chchch.github.io/sanskrit-alignment/docs/) mentions "This work is licensed under a Creative Commons Attribution 4.0 International License," while the [LICENSE file](https://github.com/chchch/sanskrit-alignment/blob/master/LICENSE) in the GitHub repo licenses the code under BSD-3. I am wondering if it would be better to pick one license?
