### 2022-02-09 #4022 @xiaohk
Hey @galessiorob, could you please add me as an assignee for this issue? Currently I can't click the checklist :P

Hi @chchch, I have finished my review. This work looks very nice and the software is well-organized. I especially like the documentation page.

My main concern is the potentially inconsistent license for this project (see https://github.com/chchch/sanskrit-alignment/issues/7). I am okay with the current auto-test set up.

One minor suggestion is to cite the papers listed in the [documentation page](https://chchch.github.io/sanskrit-alignment/docs/) in your software paper as well. It can help the authors of these papers get proper credits (e.g., the software paper will be indexed in google scholar).

Thanks!
