### 2022-02-09 #4022 @chchch
Hmm, I'm actually not so clear about the difference between licensing text and software — is it better to use a software license to cover documentation as well?
