### 2022-02-09 #4022 @xiaohk
A fix might be changing "this work" to "this documentation" in the documentation, and declare dual-license in your repo (since the documentation is also included in the repo): [Declaring multiple licences in a GitHub project](https://softwareengineering.stackexchange.com/questions/304874/declaring-multiple-licences-in-a-github-project).
