### 2022-02-12 #4022 @galessiorob
@chchch I put together a few edit suggestions in this [PR.](https://github.com/chchch/sanskrit-alignment/pull/8)
