### 2022-02-21 #4022 @galessiorob
@kinow I believe we're ready to accept the paper and publish it, would you mind checking the last two items from your list if you think they've been satisfied? 🙏

@xiaohk thanks again for your thorough review! 🎉

@chchch I think we're ready to publish! The next thing we need from you is to deposit your software in Zenodo and post the DOI here and the most recent version. After that, we can set the final version, generate a final proof, and have an Editor in Chief do the publishing. 🙏
