### 2022-02-21 #4022 @kinow
Hi @galessiorob !

>@kinow I believe we're ready to accept the paper and publish it, would you mind checking the last two items from your list if you think they've been satisfied? pray

Sure!

>>    State of the field: Do the authors describe how this software compares to other commonly-used packages?
>
>In the paper draft I believe you mention that while the techniques used in helayo is similar to those used in the bioinformatics field, but very different to modern bioinformatics techniques that are too specific to be used in natural language processing.
>
>But reading the paper I had a few questions, like “could I use it for Chinese and Japanese manuscripts?”, I believe they suffer similar issue to what you described in the motivation section of the paper. Some papers include a section about future work, improvements, etc, not sure if necessary here (cc @galessiorob ), but looks like it could be useful for developers/linguists/etc interested in your work.

Just had a brief look at the paper, and it looks OK! Nothing to comment here, good job!

>Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

The paper talks about `helayo` and also about the `matrix-editor`, the web UI.

This requirement for automated tests is still missing in `matrix-editor`. The `helayo` program has the dependencies managed via a package manager, and unit tests. But `matrix-editor` doesn't have either (https://github.com/chchch/sanskrit-alignment/issues/3).

I thought about checking that option since the paper appears to be about `helayo`, but I think `matrix-editor` is really important & useful to readers and users. The `matrix-editor` is mentioned and has figures of the `matrix-editor` in the paper, and the source code is in the same repository (it's almost as if it's part of `helayo`?)

So I think that option remains unchecked. Not sure if that's a blocker for the publication @galessiorob ? If not my review is complete.

>Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

`matrix-editor` has the dependencies files copied in the repository, instead of using an automated package management solution, but I noted that in the above it said "ideally", that's why I checked that :+1: but would be better to use yarn/npm/cnd/etc instead, IMO.

Bruno
