### 2022-02-21 chchch/sanskrit-alignment#7 @kinow
Good catch @xiaohk . I guess maybe @chchch meant that the docs/text was CC 4, not the source code? But I am not sure, so definitely better to clarify that :+1:
