### 2022-02-21 #4022 @galessiorob
@kinow that's actually a very good point...

@chchch Do you think you could address the above points about the `matrix-editor` so that the program is more robust?
