### 2022-02-21 #4022 @chchch
Hi @kinow, I actually moved to a yarn/snowpack build system like you suggested. I also did a big code cleanup and started adding tests for the matrix-editor; there's a GitHub action that runs them. There could definitely be more tests, but at least there's a simple, solid test harness now.

@galessiorob, thanks for the clarification, I'll update the paper with your suggestions and also try to add more references like @xiaohk suggested. I wasn't sure how long a JOSS paper should be — the docs seemed to suggest that I should keep it short, but I've since seen some JOSS papers that are quite long. I guess there's space to add a couple of lines and references?
