### 2022-02-21 #4022 @kinow
@chchch ah, apologies. I remember you saying you were working on that, but I saw the GH issue still open and didn't find the `package.json` in the `matrix-editor` [directory](https://github.com/chchch/sanskrit-alignment/tree/master/matrix-editor).

But just had another look and found if in the `src` directory, and also found the tests :slightly_smiling_face:

@galessiorob my review is complete, I'm checking the final checkbox. Thanks!!! :tada:
