### 2022-02-21 chchch/sanskrit-alignment#3 @kinow
@chchch I think this issue can now be closed? :tada:

Thanks!
