### 2022-02-21 chchch/sanskrit-alignment#3 @chchch
Yes thanks! Sorry I was focusing on your pull request and forgot about the issue.
