### 2022-02-22 #4022 @galessiorob
... my bad too for not looking in the `src` directory. Thank you for adding the tests, and nice use of GH Actions 🚀

As for the length, yes, don't feel restricted to add a bit more and of course, up to your discretion. Let me know when you've done all the changes you want to make and we can give it a final pass before recommending it for publishing.

Thanks for all your work so far!
