+++ paper.bib
@@ -42,6 +42,22 @@ address = {Cambridge},
 year = 1997
 }
 
+@book {KatreGode:1941,
+author = {{Katre}, Sumitra Mangesh and {Gode}, Parshuram Krishna},
+title = "Introduction to Indian Textual Criticism",
+publisher = {Karnatak Publishing House},
+address = {Bombay},
+year = 1941
+}
+
+@book {West:1973,
+author = {{West}, {Martin Litchfield}},
+title = "Textual Criticism and Editorial Technique Applicable to Greek and Latin Texts",
+publisher = {B. G. Teubner},
+address = {Stuttgart},
+year = 1973
+}
+
 @article {Fitch:1971,
 author = {{Fitch}, Walter M.},
 title = "Defining the course of Evolution: Minimum change for a specific tree topology",
