+++ paper.bib
@@ -15,7 +15,7 @@ year = 2007
 @incollection {Maas:2013,
 author = "Philipp A. Maas",
 title = {On What to Do with a Stemma -- Towards a Critical Edition of the {Carakasaṃhitā Vimānasthāna} 8.29},
-booktitle = "Medical Texts and Manuscripts in Indian Cultural History",
+booktitle = "Medical Texts and Manuscripts in {Indian} Cultural History",
 editor  = {{Wujastyk}, Dominik and {Cerulli}, Anthony and {Preisendanz}, Karin},
 publisher = {Manohar},
 address = {New Delhi},
@@ -44,7 +44,7 @@ year = 1997
 
 @book {KatreGode:1941,
 author = {{Katre}, Sumitra Mangesh and {Gode}, Parshuram Krishna},
-title = "Introduction to Indian Textual Criticism",
+title = "Introduction to {Indian} Textual Criticism",
 publisher = {Karnatak Publishing House},
 address = {Bombay},
 year = 1941
@@ -52,7 +52,7 @@ year = 1941
 
 @book {West:1973,
 author = {{West}, {Martin Litchfield}},
-title = "Textual Criticism and Editorial Technique Applicable to Greek and Latin Texts",
+title = "Textual Criticism and Editorial Technique Applicable to {Greek} and {Latin} Texts",
 publisher = {B. G. Teubner},
 address = {Stuttgart},
 year = 1973
