+++ helayo/package.yaml
@@ -1,5 +1,5 @@
 name:                helayo
-version:             0.5
+version:             0.65
 github:              "chchch/sanskrit-alignment"
 license:             GPL-3
 author:              "Charles Li"

[…]
