+++ .github/workflows/build.yml
@@ -1,62 +1,0 @@
-name: helayo build & test
-
-on:
-  push:
-    paths:
-    - 'helayo/**'
-  workflow_dispatch:
-
-jobs:
-
-  stack:
-    name: stack / ${{ matrix.os }}
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-latest, macOS-latest, windows-latest]
-        stack: [latest]
-        ghc: ["8.10.3"]
-
-    steps:
-    - uses: actions/checkout@v2
-
-    - uses: haskell/actions/setup@v1
-      name: Setup Haskell Stack
-      with:
-        ghc-version: ${{ matrix.ghc }}
-        enable-stack: true
-        stack-version: ${{ matrix.stack }}
-
-    - uses: actions/cache@v2.1.3
-      name: Cache ~/.stack
-      with:
-        path: ~/.stack
-        key: ${{ runner.os }}-${{ matrix.ghc }}-stack
-
-    - name: Install dependencies
-      run: |
-        cd helayo
-        stack --resolver lts-17.2 build --system-ghc --test --bench --no-run-tests --no-run-benchmarks --only-dependencies
-
-    - name: Build
-      run: |
-        cd helayo
-        stack --resolver lts-17.2 build --system-ghc --test --bench --no-run-tests --no-run-benchmarks
-
-    - name: Test
-      run: |
-        cd helayo
-        stack --resolver lts-17.2 test --system-ghc
-    
-    - name: Install
-      run: |
-        cd helayo
-        git pull
-        stack --resolver lts-17.2 install --local-bin-path ./dist/${{ matrix.os }}
-
-    - name: Commit
-      uses: stefanzweifel/git-auto-commit-action@v4
-      with:
-        commit_message: build ${{ matrix.os }} binary
-        file_pattern: helayo/dist/${{ matrix.os }}
-        commit_user_name: Helayo Bot
