+++ .github/workflows/build.yml
@@ -34,10 +34,10 @@ jobs:
         key: ${{ runner.os }}-${{ matrix.ghc }}-stack
 
     # https://github.com/commercialhaskell/stack/issues/5405
-    - name: clear macOS cache
+    - name: Clear macOS cache
       if: runner.os == 'macOS'
       continue-on-error: true
-      run: rm -r ~/.stack-setup-exe-cache
+      run: rm -r ~/.stack/setup-exe-cache
 
     - name: Install dependencies
       run: |
