### 2022-03-07 #4022 @chchch
Hello,
I've made a new release so that it gets tracked by Zenodo: https://zenodo.org/account/settings/github/repository/chchch/sanskrit-alignment

I don't have a citation.cff file in the repository yet, but I guess I would do that after publication of the JOSS paper? The only thing is that the Zenodo DOI would be pointing to a release without a citation.cff file.
