### 2022-03-08 #4022 @galessiorob
[Note the missing citation.cff file](https://github.com/openjournals/joss-reviews/issues/4022#issuecomment-1060428253), please advice.
