### 2022-03-11 #4022 @kyleniemeyer
Hi @chchch can you clean up the Zenodo metadata so that the author name matches the paper? Also, it looks like the Barbook et al. reference has a DOI (10.1038/29667), could you add this? Once you resolve those things I can accept the submission.
