### 2022-03-19 #4022 @chchch
Hello,

I've made those two changes (sorry for the delay; it took a while to figure out the best way to change the Zenodo metadata). It looks like this now: https://zenodo.org/record/6334073

Let me know if that looks right!
