### 2022-03-21 #4022 @danielskatz
@chchch - here's what I see

I expect to see a title that matched that of the paper: "helayo: Reconstructing Sanskrit texts from manuscript witnesses".

The section labeled "How can I edit the metadata of a published record?" in https://help.zenodo.org may be helpful.
