### 2022-03-21 #4022 @danielskatz
👋 @kinow (and @galessiorob) - I see one checkbox is not checked (performance) - can you check this?  If not, why not?
