### 2022-03-21 #4022 @kinow
> 👋 @kinow (and @galessiorob) - I see one checkbox is not checked (performance) - can you check this? If not, why not?

I thought I had checked off that option, sorry. Done 👍
